requirements.txt
asgiref==3.8.1
Django==4.2.11
flake8==6.1.0
greenlet==3.1.1
mccabe==0.7.0
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pyflakes==3.1.0
pyModbusTCP==0.1.10
python-dotenv==1.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
typing_extensions==4.12.2
